Java Project
I did this project in a team consisting of 4 people. I made the homepage, log-in, Sign up, and Profile forms using Scene Builder in Java and OOP basics for classes. This project has many issues and null exceptions and we agreed as a team that we will complete it later in the summer after the college exams.